feat(phone): AVO-036 review fixes - phase 1
Fix team row visibility bug (always show team row for null team/assignee).
Remove Show all types picker (agent-only types not in PhoneTicketType).
Extract NoSelectTextField to shared public widget (4 copies → 2).
Cache teams Future in CreateTicketScreen.initState.
Fix flutter analyze warnings (DropdownButtonFormField, doc comments).
