+++ phone/lib/widgets/text_input_sheet.dart
@@ -1,74 +1,6 @@
 import 'package:flutter/material.dart';
 
-/// A TextField that places cursor at tap position on single tap,
-/// without selecting text. Preserves double-tap word selection.
-///
-/// Works around flutter/flutter#98720, #105185 where single taps in
-/// TextFields can unexpectedly select words instead of placing cursor.
-class _NoSelectTextField extends StatefulWidget {
-  final TextEditingController controller;
-  final InputDecoration? decoration;
-  final int? maxLines;
-  final TextInputAction? textInputAction;
-  final ValueChanged<String>? onSubmitted;
-  final bool autofocus;
-
-  const _NoSelectTextField({
-    required this.controller,
-    this.decoration,
-    this.maxLines,
-    this.textInputAction,
-    this.onSubmitted,
-    this.autofocus = false,
-  });
-
-  @override
-  State<_NoSelectTextField> createState() => _NoSelectTextFieldState();
-}
-
-class _NoSelectTextFieldState extends State<_NoSelectTextField> {
-  Offset? _tapPosition;
-
-  void _handleTapDown(TapDownDetails details) {
-    _tapPosition = details.globalPosition;
-  }
-
-  void _handleTap() {
-    if (_tapPosition == null) return;
-    final renderBox = context.findRenderObject() as RenderBox;
-    final localPosition = renderBox.globalToLocal(_tapPosition!);
-
-    // Calculate cursor offset from tap position
-    final textPainter = TextPainter(
-      text: TextSpan(text: widget.controller.text),
-      textDirection: TextDirection.ltr,
-    );
-    textPainter.layout(maxWidth: renderBox.size.width);
-
-    // Find the character offset nearest to tap position
-    final textPosition = textPainter.getPositionForOffset(localPosition);
-    widget.controller.selection = TextSelection.collapsed(
-      offset: textPosition.offset,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTapDown: _handleTapDown,
-      onTap: _handleTap,
-      behavior: HitTestBehavior.opaque,
-      child: TextField(
-        controller: widget.controller,
-        autofocus: widget.autofocus,
-        maxLines: widget.maxLines,
-        decoration: widget.decoration,
-        textInputAction: widget.textInputAction,
-        onSubmitted: widget.onSubmitted,
-      ),
-    );
-  }
-}
+import 'no_select_text_field.dart';
 
 /// A reusable bottom sheet for text input with confirm/cancel behavior.
 ///
@@ -144,7 +76,7 @@ class _TextInputSheetState extends State<TextInputSheet> {
                   ?.copyWith(fontWeight: FontWeight.w600),
             ),
             const SizedBox(height: 12),
-            _NoSelectTextField(
+            NoSelectTextFieldRaw(
               controller: _controller,
               autofocus: true,
               maxLines: null,
